Kill killswitch for events that come in via message_echoes (#37)
## Why are we doing this?

To enable two bots one page for a new conversation logger, you need the ability to make a bot that can listen to echo messages. Otherwise, the observer bot can only observe humans.

After talking through #34, the double opt-in seemed excessive. This PR solves the same problem, but with only one opt-in, and so less code.

closes #29

## Did you document your work?

README now talks about echo messages

## How can someone test these changes?

Steps to manually verify the change:

1.  Edit the webhook for a bot to get echos too
2. view the log to see more events come in

## What possible risks or adverse effects are there?

* none. If a user gets echos and it triggers things, then the app should either stop listening to echos and/or they should check the echo themselves.

## What are the follow-up tasks?

* none

## Are there any known issues?

none

## Did the test coverage decrease?

stays the same. code was deleted, and that code wasn't really covered, so... maybe this increases coverage?