Add printf-like formatting to Text response helper (#42)
## Why are we doing this?

The text translation helps us write smaller code, but if you need to do special formatting, like "Fetched %d out of %d", you currently have to write out the whole string yourself, which can lead to very long lines, which is one of the major reasons why the `Text` translation layer exists in the first place. This adds [`util.format`](https://nodejs.org/docs/latest/api/util.html#util_util_format_format_args) to `Text` so it knows how to do this kind of formatting in a way that's consistent with `console` and printf and gettext.

closes #22

## Did you document your work?

New functionality is in the README and there are test cases for some edge cases where I wondered what would happen.

## How can someone test these changes?

Steps to manually verify the change:

Since this is in the SDK, there's only unit tests unless you `npm link` in and do a manual test. I trusted the test cases.

2. `npm i`
3. `npm t`

## What possible risks or adverse effects are there?

There is a computation cost in adding `util.format`. It may also be unnatural when `%` naturally occurs in your text

## What are the follow-up tasks?

* roll a release

## Are there any known issues?

none

## Did the test coverage decrease?

stays the same