Document session.profile (#43)
## Why are we doing this?

This is a documentation update for some missing information. I tried looking for it in the past assuming it was there. Since `session.profile` is provided by the SDK, it's worth writing something even if it does just point to Facebook's docs.

closes #32 

## Did you document your work?

this is it

## How can someone test these changes?

Steps to manually verify the change:

1. view the [rendered readme](https://github.com/CondeNast/launch-vehicle-fbm/blob/d9fc92ec34948b257bb9699613ddc2411b188c7f/README.md)
2. clicking on a hotlinked "a session object" should take you to the session docs

## What possible risks or adverse effects are there?

* increases docs tech debt, but I managed to cop out by trying to get people to go to the official docs for more info

## What are the follow-up tasks?

none

## Are there any known issues?

N/A

## Did the test coverage decrease?

N/A