Store the page id for conversations in the session (#44)
## Why are we doing this?

The long term goal is to have one bot service multiple pages. To do that, we need to associate conversations with a page. Lucky for us, we have the `session` storage around. This PR:
* adds an internal `_pageId` to the session object
* adds some Flow type documentation I wish existed to document some function signatures
* fixes some leaky tests

This is a precursor to #30

## Did you document your work?

The session object doesn't require a README update because it's internal. There are new Flow docs. Functionality and the API does not change.

## How can someone test these changes?

Just unit tests. I also ran this `npm link`ed to another bot and it ran as usual.

## What possible risks or adverse effects are there?

* minor: increases the payload and responsibility of the session object

## What are the follow-up tasks?

* #30 Sending based on page id
* major rework of page id/page access token config

## Are there any known issues?

none

## Did the test coverage decrease?

new code is covered, 
increases with flow coverage
improves exiting tests