Add ability to bring your own cache storage (#48)
## Why are we doing this?

We need to expand the session cache to live outside the current memory cache. This PR adjusts the `Messenger` constructor to take in a new optional option: `cache`. If supplied, `Messenger` will use this new cache instead of the default.

Some other changes:
* refactor to reduce Flow type warnings
* refactor to remove `__internals__` hack
* remove unnecessary namespacing that the redis cache backend handles already

closes #47 
closes #41 (makes this obsolete)
closes #45 (documentation)

## Did you document your work?

The readme documents the basics of how to supply your own cache. If the cacheman-redis library were more elegant, I'd include an example but it's very awkward to use so I opted to punt on including an example to avoid including a confusing example.

## How can someone test these changes?

Steps to manually verify the change:

1. `npm i`
2. `npm t`
3. you can also `npm link` this and verify that your bot behaves the same.

## What possible risks or adverse effects are there?

* none. This does open another way for the user to shoot themselves in the foot though.

## What are the follow-up tasks?

* document an example. Perhaps just a dummy cache example?

## Are there any known issues?

None

## Did the test coverage decrease?

There's a weird behavior the redis cache does where it returns `null` which then throws off the default parameters. There's code added to workaround when a cache backend does this along with tests.