Postback check for 'greetings' & 'help' and payload consistency for postbacks and quick replies (#52)
This PR ties up some loose ends on how postbacks are handled by checking the postback payload against the _greeting_ and _help_ regular expressions if enabled. Persistent menu clicks are emitted as postbacks and the SDK was not catching requests for _help_ that originated from a persistent menu.

Additionally, while `text` events normalized the user provided content by trimming whitespace and converting to all lower case; postbacks and quick replies were emitting payloads (as `text`) without this normalization. This PR emits the `text` in postback and quick replies as normalized text and adds a `payload` attribute that is the unmodified payload.