Updating dependencies, new lint rules, satisfying new lint rules (#68)
The `eslint-config-condenast` package was recently updated, reducing the difference between it and `eslint-config-airbnb`, the package on which it is based. For that reason, I sought to run `npm-check` on this repo and update the dependencies. The `eslint` config was also updated/simplified to reflect this and as a result, a new dev dependency, `eslint-plugin-mocha` was added.

Lastly, code changes - mostly whitespace and `function` notation updates - were completed to comply with the new lint rules.