Bump cache timeout from 1hr to 1day (#70)
## Why are we doing this?

While doing prep work for a feature, I decided to change the defaults to make them more realistic based on our usage of the SDK. By changing this value, users will get more value out of the bot using just the default configuration. We previously bumped it from 1 minute (the default for the base library) to 1 hour. But when users tested bots, they would often leave for over an hour and come back.

To get users to switch to a real cache end, there's a new line to go look in the wiki for a Redis example. There isn't a great example right now, but if they look at the multitenant example, they'll find Redis.

## Did you document your work?

yep

## How can someone test these changes?

Steps to manually verify the change:

1. npm i
2. npm t

## What possible risks or adverse effects are there?

* More memory usage, but just barely
* something breaking in a dependency

## What are the follow-up tasks?

* make a more contrived Redis example in the wiki

## Are there any known issues?

none

## Did the test coverage decrease?

n/a