Add "pause" webhook to allow for live person takeovers (#71)
# Why are we doing this?

For most bots, users will want the ability to pause the bot just for a user so they can talk to them without the bot's automated responses filling the conversation. An example would be a customer service bot where the user signals they're ready to talk to a human. This adds a webhook that sets a cache key that will cause the bot to stop processing just for a user.

Follow Dashbot's guide: https://www.dashbot.io/sdk/pause

closes #67 

## Did you document your work?

I added a new section to the README for the http endpoints available.

## How can someone test these changes?

Steps to manually verify the change:

Just unit tests since this is the sdk.

1. `npm i`
2. `npm t`

## What possible risks or adverse effects are there?

* there's a new cache key, so that means you need another cache lookup, which increases the latency to the message

## What are the follow-up tasks?

* none

## Are there any known issues?

none

## Did the test coverage decrease?

new code is covered