Add support for Text quick replies (#75)
## Why are we doing this?

We made a way to do Image + quick replies but not Text + quick replies. This adds that missing message type, it also deprecates the old `ImageQuickReply` helper in favor of plain `Text` and `Image` + `.quickReplies()` method. Adding quick replies isn't really a new message type, it's a modification of existing types. The syntax from the tests and README look like:

    const text = new Text('Where are you?').quickReplies([{ content_type: 'location' }])

closes #60 

## Did you document your work?

README is updated. I wonder if this should move to the wiki though, the README is getting heavy.

## How can someone test these changes?

Steps to manually verify the change:

1. `npm i`
2. `npm t`

## What possible risks or adverse effects are there?

* none

tech debt increases slightly but can improve when `ImageQuickReply` is removed. There's also some potential for a DRY refactor

## What are the follow-up tasks?

* release minor version bump

## Are there any known issues?

none

## Did the test coverage decrease?

new code is covered, increasing coverage