Use fallback instead of exploding when failing to get a profile (#77)
## Why are we doing this?

The bot tries to get a profile, but it's a convenience, not a requirement. If for some reason the profile can't be loaded, it should not be a show stopper. There are already cases where we can't get a profile (bot messing a bot) where we fallback to an empty object and log a warning.

This PR also implements a longstanding related TODO and fixes the fallback case for the greeting message.

## Did you document your work?

No README updates necessary. Code comments and type defs are updated.

## How can someone test these changes?

Steps to manually verify the change:

1. `npm t`

## What possible risks or adverse effects are there?

* none. If someone was relying on this obscure code blowing up, that should be fixed on their side

## What are the follow-up tasks?

* none

## Are there any known issues?

none

## Did the test coverage decrease?

increases, the profile fetching logic inside `routeEachMessage` wasn't covered before.