Add "Quick Start" to Sphinx docs (#81)
## Why are we doing this?

This expands the Sphinx documentation to add a "Quick start" section. The new doc was then used to write https://github.com/CondeNast/rkt-tv-summit-bot/pull/1

## Did you document your work?

dis is it

## How can someone test these changes?

I didn't do a ReadTheDocs build, but you can examine the new restructured text files within GitHub.

## What possible risks or adverse effects are there?

This begins the breaking up of docs over multiple pages. Personally, I like keeping things to one page, but if you're sticking to one page, there's not really much point to using Sphinx, the README is better for that, and we outgrew that already.

## What are the follow-up tasks?

* keep writing docs

## Are there any known issues?

none

## Did the test coverage decrease?

N/A