Enable 'referral' event for Messenger Code scans (#82)
## Why are we doing this?

Messenger has "parametric messenger codes". When these visual codes are scanned, the bot is supposed to be able to respond. Facebook created a new `referral` event for this, even though it's the same thing as `postback` 🤷‍♂️ . This PR adds support for this new event, following the naming conventions for the other events.

https://developers.facebook.com/docs/messenger-platform/discovery/messenger-codes/

## Did you document your work?

README has the new event and how to use it.

## How can someone test these changes?

Steps to manually verify the change:

1. `npm i`
1. `npm t`

## What possible risks or adverse effects are there?

* none

## What are the follow-up tasks?

* roll a new minor release
* migrate tests to use real data (todo in the codebase)

## Are there any known issues?

none

## Did the test coverage decrease?

new code is covered, albeit not quite unit level, not quite integration level, but tests with this pattern should be easier to maintain.