Split up debug message to be more developer friendly. (#89)
Why are we doing this?

Per my comment on #63, a debug message referencing serverUrl is what throws this error message. Within the SDK, serverUrl is only used for Login with Facebook support and this debug message.

To remove this as a blocker for developers, I split up the debug message and log an error if no serverUrl is present. This should be enough of an alert to a developer without throwing a blocking exception that is not helpful to their development.

I also ran npm-check to update dependencies and made some tweaks per the lint rules.

Did you document your work?
No README updates necessary.

How can someone test these changes?
Run a simple example bot before and after and note the change in results. The master branch of the enterprise-fbm repo can be used if desired.

What possible risks or adverse effects are there?
The Login with Facebook portions of the SDK will not work properly without a serverUrl. It is possible that with this non-blocking error message that a developer could overlook the need for the configuration item. That said, this is a lesser used part of the SDK, so the risk seems low. It would also be apparent to a developer that something is wrong when they attempt to run through the login flow during their testing.

What are the follow-up tasks?
None identified.

Are there any known issues?
None identified.

Did the test coverage decrease?
No.